python/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Clean multi-lang AI agents"
dependencies = [
    "ruff",
    "pre-commit",
    "google-adk",
    "mypy>=1.19.1",
    "pytz>=2026.1.post1",
    "types-pytz>=2026.1.1.20260304",
    "python-dotenv>=1.2.2",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "litellm>=1.82.6",
    "pyright>=1.1.408",
    "toolbox-adk>=0.7.0",
    "google-auth>=2.49.1",
    "mcp>=1.26.0",
    "httpx>=0.28.1",
]
requires-python = ">=3.10"
